Extract NetworkSwitcher
